servers/mcp-deepcell/pyproject.toml
[project]
name = "mcp-deepcell"
version = "0.2.0"
description = "MCP server for deep learning cell segmentation with DeepCell-TF"
authors = [{name = "Precision Medicine MCP Team"}]
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "scikit-image>=0.22.0",
    # DeepCell Integration (Phase 1)
    # Note: deepcell requires tensorflow 2.8.x
    "deepcell>=0.12.0",
    "scipy>=1.11.0",
]

[project.urls]
Homepage = "https://github.com/lynnlangit/precision-medicine-mcp"
Issues = "https://github.com/lynnlangit/precision-medicine-mcp/issues"
Documentation = "https://github.com/lynnlangit/precision-medicine-mcp/tree/main/docs"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "black>=24.0.0"]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
